Run jupyter lite with subprocess.run to not suppress stdout
check_output redirects stdout, so we don't see it when we do sphinx-build. We instead switch to using subprocess.run so that we can see the stdout output.

Suppressing stdout was causing errors in jupyter lite to not be displayed, which was very confusing.

A separate issue is that jupyter lite is not returning a nonzero exit code when there is an error. But that is a problem to solve in jupyter lite. Once it is solved, having the check=True means that the sphinx build will also raise an error.