Misc css and documentation update.
Here are a couple of modification:

The example_class was not applied to all the container.

Having both the default and iframe container with the same class was a
bit annoying for some css, and lead to `.hidden` not being applied in some
of my experiments, so swap the name of the iframe container.

Add an example in the documentation on how to change the color and
positioning of the button.

I happen to have prettier install for css files, so it also reformatted
the css to use css standard of 2 spaces.

Move the main colors into css variables, so that if need and if we start
to use those colors in more places they are more easily overwritten.

Also contains mist parsing warnings.

 - Unknown escape sequence \;, need double escape.
 - pygments say yml is unknown but yaml works