Suggestion: Add pre-commit to format js and css files.
I don't believe most of us know the current conventions and have the
tooling installed. So it might be a good thing to have a consistent
style.

If we do this we can also add pre-commit-ci so that PR are automatically
updated. I'm generally not a fan of pre-commit hooks in general and
pre commit-ci as it either can take forcer before you can already
commit, or also when CI push on the PR, prevent pushing before pulling
which can be annoying for contributor. So either choice is good for me.