Run pre-commit on all files in this repository.
This should only reformat JS and css files.

I've requested precommit-ci to be insalled on the JupyterLIght org, but
do not have enough permission to allow it.

One it is done and configured, it will automatically update PRs by
autoformatting, so a first pass to autoformat everything is good to not
have noise in later PRs.